updating notes

+++ Chapter02/math.cpp
@@ -4,7 +4,7 @@
 using namespace std;
 
 int main() {
-  double totalDistance;  // result
+  double totalDistance;  // result??
 
   // point 1
   double x1{2.4};
@@ -14,16 +14,11 @@ int main() {
   double y2{10.0};
 
   // TODO: Implement solution
-  double xDist = pow(x2 - x1, 2);
-  double yDist = pow(y2 - y1, 2);
-
+  int xDist = pow(x2 - x1, 2);
+  int yDist = pow(y2 - y1, 2);
   totalDistance = sqrt(xDist + yDist);
 
   cout << "Distance between points: " << totalDistance << endl;
 
   return 0;
 }
-
-// double xDist = pow(x2 - x1, 2);
-// double yDist = pow(y2 - y1, 2);
-// totalDist = sqrt(xDist + yDist);
